@@ -16,11 +16,11 @@
         }
     ],
     "require": {
-        "php": "^7.1",
-        "illuminate/session": "^5.5|^5.6|^5.7|^5.8|^6.0"
+        "php": "^7.4",
+        "illuminate/session": "^5.8|^6.0"
     },
     "require-dev": {
-        "orchestra/testbench": "^3.5|^3.6|^3.7|^3.8|^4.0"
+        "orchestra/testbench": "^3.8|^4.0"
     },
     "autoload": {
         "psr-4": {
